requirements.txt
httpx==0.28.1
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-experimental==0.3.4
langchain-openai==0.2.14
openai==1.58.1
openpyxl==3.1.5
pandas==1.5.3
pandasai==2.4.1
pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
seaborn==0.13.2
streamlit==1.41.1
tabulate==0.9.0